Update Monte-carlo-simulation-generate_v1.0.py
A python code for simulating atmospheric muons distribution below sea level for the KM3Net experiment. starting from an analytically fitted formula at sea level.
Then the muon is propagated through water, while taking into account energy loss and the geometrical acceptance of the detector.
some geometrical optimization were added to reduce the run time of the code.
Additional the code is parallelized to run on multiple cpu's since the simulation is computationally demanding. A maximum of 14 cores is put in the code. If you need 
to use more, modify this number.
the code was written during an inernship in the LPC lab, Caen, France, 2020.
note: the execution requires the support file named: loss.txt